refactor: extract config hash logic to new file
